Fixed bug with swipes, tap and slow processing
The main problems was not in the header file. It was system("cls"), situated in while in main function. It works really slow and because of it RSGManager skipped some frames and because of it not recognized some gestures. It is simply information about bug and hotfix on application. I'll try to separate recognition process from main process by creating a parallel thread. It must solve the problem.